+++ src/client/server_interface.hpp
@@ -23,7 +23,7 @@ public:
 
 	virtual void toggleFly() = 0;
 	virtual void setPlayerMoveInput(int moveInput) = 0;
-	virtual void setPlayerOrientation(double yaw, double pitch) = 0;
+	virtual void setPlayerOrientation(float yaw, float pitch) = 0;
 	virtual void setBlock(uint8 block) = 0;
 
 	virtual void edit(vec3i64 block, uint8 type) = 0;
