+++ src/io/archive.cpp
@@ -95,7 +95,7 @@ void ArchiveFile::initialize() {
 	header.endianess_bytes = ENDIANESS_BYTES;
 	_endianess = NATIVE;
 	_version = header.version = 0x0001;
-	size_t size = _region_size * _region_size * _region_size;
+	uint size = _region_size * _region_size * _region_size;
 	header.size = size;
 	_directory_offset = header.directory_offset = sizeof (Header);
 
@@ -131,9 +131,9 @@ bool ArchiveFile::loadChunk(Chunk &chunk) {
 
 bool ArchiveFile::loadChunk(vec3i64 cc, Chunk &chunk) {
 	// read directory entry
-	int64 x = cycle(cc[0], _region_size);
-	int64 y = cycle(cc[1], _region_size);
-	int64 z = cycle(cc[2], _region_size);
+	size_t x = cycle(cc[0], _region_size);
+	size_t y = cycle(cc[1], _region_size);
+	size_t z = cycle(cc[2], _region_size);
 	size_t id = x + (_region_size * (y + (_region_size * z)));
 	_file.seekg(_directory_offset + id * sizeof (DirectoryEntry));
 	DirectoryEntry dir_entry;
@@ -177,9 +177,9 @@ bool ArchiveFile::loadChunk(vec3i64 cc, Chunk &chunk) {
 void ArchiveFile::storeChunk(const Chunk &chunk) {
 	// read directory entry
 	vec3i64 cc = chunk.getCC();
-	int64 x = cycle(cc[0], _region_size);
-	int64 y = cycle(cc[1], _region_size);
-	int64 z = cycle(cc[2], _region_size);
+	size_t x = cycle(cc[0], _region_size);
+	size_t y = cycle(cc[1], _region_size);
+	size_t z = cycle(cc[2], _region_size);
 	size_t id = x + (_region_size * (y + (_region_size * z)));
 	_file.seekg(_directory_offset + id * sizeof (DirectoryEntry));
 	DirectoryEntry dir_entry;
@@ -234,7 +234,7 @@ void ArchiveFile::storeChunk(const Chunk &chunk) {
 	} else {
 		// chunk did not already exist or is now too big for the old space
 		_file.seekg(0, ios_base::end);
-		dir_entry.offset = _file.tellg();
+		dir_entry.offset = (uint32) (_file.tellg());
 		dir_entry.size = store_size;
 		_file.seekp(_directory_offset + id * sizeof (DirectoryEntry));
 		_file.write((char *)(&dir_entry), sizeof (DirectoryEntry));
